+++ pkgs/ffigen/test/native_objc_test/global_test.dart
@@ -37,12 +37,12 @@ void main() {
     (Pointer<ObjCObject>, Pointer<ObjCObject>) globalObjectRefCountingInner() {
       final obj1 = NSObject.new1();
       lib.globalObject = obj1;
-      final obj1raw = obj1.pointer;
+      final obj1raw = obj1.ref.pointer;
       expect(objectRetainCount(obj1raw), 2); // obj1, and the global variable.
 
       final obj2 = NSObject.new1();
       lib.globalObject = obj2;
-      final obj2raw = obj2.pointer;
+      final obj2raw = obj2.ref.pointer;
       expect(objectRetainCount(obj2raw), 2); // obj2, and the global variable.
       expect(objectRetainCount(obj1raw), 1); // Just obj1.
       expect(obj1, isNotNull); // Force obj1 to stay in scope.
@@ -74,12 +74,12 @@ void main() {
         globalBlockRefCountingInner() {
       final blk1 = ObjCBlock_Int32_Int32.fromFunction((int x) => x * 10);
       lib.globalBlock = blk1;
-      final blk1raw = blk1.pointer;
+      final blk1raw = blk1.ref.pointer;
       expect(blockRetainCount(blk1raw), 2); // blk1, and the global variable.
 
       final blk2 = ObjCBlock_Int32_Int32.fromFunction((int x) => x + 1000);
       lib.globalBlock = blk2;
-      final blk2raw = blk2.pointer;
+      final blk2raw = blk2.ref.pointer;
       expect(blockRetainCount(blk2raw), 2); // blk2, and the global variable.
       expect(blockRetainCount(blk1raw), 1); // Just blk1.
       expect(blk1, isNotNull); // Force blk1 to stay in scope.