+++ pkgs/objective_c/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.2.0-wip
+## 2.0.0-wip
 
 - Drop API methods that are deprecated in the oldest versions of iOS and macOS
   that flutter supports.
@@ -7,6 +7,10 @@
 - Enable ObjC objects and blocks to be sent between isolates.
 - Add `autorelease` and `retainAndAutorelease` methods to ObjC objects and
   blocks.
+- __Breaking change__: Remove some convenience methods from `_ObjCRefHolder`:
+  `isReleased`, `release`, `pointer`, and `retainAndReturnPointer`. Uses of
+  these methods now need to go through `.ref`. Eg `obj.pointer` becomes
+  `obj.ref.pointer`.
 
 ## 1.1.0